+++ packages/react-html/src/server/components/tests/Script.test.tsx
@@ -16,10 +16,24 @@ describe('<Script />', () => {
 
     expect(script).toContainReactComponent('script', {
       src: 'foo.js',
-      type: 'text/javascript',
       integrity: '00000000',
       crossOrigin: 'anonymous',
       defer: true,
+    });
+  });
+
+  it('defaults to text/javascript', () => {
+    const script = mount(<Script src="foo.js" />);
+
+    expect(script).toContainReactComponent('script', {
+      type: 'text/javascript',
+    });
+  });
+
+  it('renders provided types', () => {
+    const moduleScript = mount(<Script src="foo.js" type="module" />);
+    expect(moduleScript).toContainReactComponent('script', {
+      type: 'module',
     });
   });
 });