+++ packages/react-html/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Better ES module support via `<Html />`'s `script` / `blockingScripts` accepting a `type` property
+
 ## [10.1.0] - 2020-12-18
 
 ### Added