+++ packages/react-html/CHANGELOG.md
@@ -7,8 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-- Minor: Only render the last meta tag if they have same name or property []()
-- Minor: Cleanup tags from DOM when unmounting []()
+- Minor: Only render the last meta tag if they have same name or property [#1732](https://github.com/Shopify/quilt/pull/1732)
+- Patch: Cleanup tags from DOM when unmounting [#1732](https://github.com/Shopify/quilt/pull/1732)
 
 ### Added