+++ packages/react-html/CHANGELOG.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [10.2.1] - 2021-02-10
 
 ### Fixed
 
-- Remove duplicated metas without mutating the state [#1736](https://github.com/Shopify/quilt/pull/1736)
+- Remove duplicated metas without mutating the state [#1736](https://github.com/Shopify/quilt/ pull/1736)
 
 ## [10.2.0] - 2021-02-03